Move clients to dedicated folder
