+++ src/infrastructure/clients/sendgrid_email_client.go
@@ -1,4 +1,4 @@
-package infrastructure
+package clients
 
 import (
 	"fmt"
